feat(docker): introduce StopCommand trait for stopping containers
Introduce the StopCommand trait to encapsulate the logic for stopping
Docker containers, enhancing modularity and reusability. This change
removes the stop method from DockerClient and integrates the trait,
streamlining the codebase. The DockerStopOutput class is updated to
handle ContainerId objects, improving type safety. Additionally, a
new StopCommandTest is added to ensure the functionality of the
StopCommand trait, while related tests are removed from
DockerClientTest to avoid redundancy. This refactor improves code
organization and test coverage.
